requirements.txt
selenium == 4.33.0
pandas == 2.2.3
python-dotenv == 1.1.0
